Set sensible initial sizes for growing arrays
Any values should work here, and it doesn't seem to make a siginificant
difference to startup time, as this only adds or reduces a few realloc
calls.

Starts input buffer at 4k of memory, which is nice for being the size of
a page and the size usually used in the read syscall. Initial choice
capacity set to 128, by assuming that candidates average 32 characters
long.