[RFC] tty_interface: If backspace is pressed on empty input, exit
This is how the : prompt in vim handles backspace on empty input, so it
feels convenient for me.

This patch may be bad for people who press backspace repeatedly to clear
the input line, and then accidentally press backspace once too often,
and exit fzy, so maybe it should be placed behind a command line option.
(I use ^U when I want to clear the input line, so this doesn't happen to me.)