Add Note About OS X Command
While I was installing this on my OS X High Sierra installation, I ran
into an issue with the `find` command. On OS X you need to provide
`find` with a path.
I'm sure this is something that devs would eventually figure out on
their own, but it I feel is could stand to be in the README too.