Prefer trailing matches over start of words
Count a match at the beginning and end of the candidate as a consecutive
match. With that fzy returns better results, when searching for file
extensions, e.g. when searching for "appc", it scores "src/app.c" higher
than "build/app/contents/cache".

A match at the beginning was scored with the same bonus as a character
after a slash. To keep it consistent, this is now scored as a
consecutive match.